Moved nav into side nav and conditionally rendered in app. Moved flex container of dashboard up into app to maintain responsiveness across all components
